infra/airflow/requirements.txt
apache-airflow-providers-postgres==5.11.0
apache-airflow-providers-http==4.10.0
dbt-postgres==1.8.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
pyyaml>=6.0
requests>=2.31
